+++ functions/npm-diff/validate.js
